requirements.txt
PyYAML
numpy
requests
pandas
tqdm